luci-app-openclash upgrade to v0.46.064

+++ package/luci-app-openclash/luasrc/controller/openclash.lua
@@ -32,10 +32,6 @@ function index()
 	entry({"admin", "services", "openclash", "coreupdate"},call("action_coreupdate"))
 	entry({"admin", "services", "openclash", "flush_fakeip_cache"}, call("action_flush_fakeip_cache"))
 	entry({"admin", "services", "openclash", "download_rule"}, call("action_download_rule"))
-	entry({"admin", "services", "openclash", "download_netflix_domains"}, call("action_download_netflix_domains"))
-	entry({"admin", "services", "openclash", "download_disney_domains"}, call("action_download_disney_domains"))
-	entry({"admin", "services", "openclash", "catch_netflix_domains"}, call("action_catch_netflix_domains"))
-	entry({"admin", "services", "openclash", "write_netflix_domains"}, call("action_write_netflix_domains"))
 	entry({"admin", "services", "openclash", "restore"}, call("action_restore_config"))
 	entry({"admin", "services", "openclash", "backup"}, call("action_backup"))
 	entry({"admin", "services", "openclash", "backup_ex_core"}, call("action_backup_ex_core"))
@@ -103,18 +99,19 @@ local json = require "luci.jsonc"
 local uci = require("luci.model.uci").cursor()
 local datatype = require "luci.cbi.datatypes"
 local opkg
+local device_name = uci:get("system", "@system[0]", "hostname")
+local device_arh = luci.sys.exec("uname -m |tr -d '\n'")
+
 if pcall(require, "luci.model.ipkg") then
 	opkg = require "luci.model.ipkg"
+else
+	opkg = nil
 end
 
 local core_path_mode = uci:get("openclash", "config", "small_flash_memory")
 if core_path_mode ~= "1" then
-	dev_core_path="/etc/openclash/core/clash"
-	tun_core_path="/etc/openclash/core/clash_tun"
 	meta_core_path="/etc/openclash/core/clash_meta"
 else
-	dev_core_path="/tmp/etc/openclash/core/clash"
-	tun_core_path="/tmp/etc/openclash/core/clash_tun"
 	meta_core_path="/tmp/etc/openclash/core/clash_meta"
 end
 
@@ -124,10 +121,6 @@ end
 
 local function is_web()
 	return luci.sys.call("pidof clash >/dev/null") == 0
-end
-
-local function restricted_mode()
-	return uci:get("openclash", "config", "restricted_mode")
 end
 
 local function is_watchdog()
@@ -158,14 +151,6 @@ local function lhie1()
 	return os.date("%Y-%m-%d %H:%M:%S",fs.mtime("/usr/share/openclash/res/lhie1.yaml"))
 end
 
-local function ConnersHua()
-	return os.date("%Y-%m-%d %H:%M:%S",fs.mtime("/usr/share/openclash/res/ConnersHua.yaml"))
-end
-
-local function ConnersHua_return()
-	return os.date("%Y-%m-%d %H:%M:%S",fs.mtime("/usr/share/openclash/res/ConnersHua_return.yaml"))
-end
-
 local function chnroute()
 	return os.date("%Y-%m-%d %H:%M:%S",fs.mtime("/etc/openclash/china_ip_route.ipset"))
 end
@@ -175,19 +160,8 @@ local function chnroutev6()
 end
 
 local function daip()
-	local daip, lan_int_name
-	lan_int_name = uci:get("openclash", "config", "lan_interface_name") or "0"
-	if lan_int_name == "0" then
-		daip = luci.sys.exec("uci -q get network.lan.ipaddr |awk -F '/' '{print $1}' 2>/dev/null |tr -d '\n'")
-	else
-		daip = luci.sys.exec(string.format("ip address show %s | grep -w 'inet'  2>/dev/null |grep -Eo 'inet [0-9\.]+' | awk '{print $2}' | tr -d '\n'", lan_int_name))
-	end
-	if not daip or daip == "" then
-		daip = luci.sys.exec("ip address show $(uci -q -p /tmp/state get network.lan.ifname || uci -q -p /tmp/state get network.lan.device) | grep -w 'inet'  2>/dev/null |grep -Eo 'inet [0-9\.]+' | awk '{print $2}' | tr -d '\n'")
-	end
-	if not daip or daip == "" then
-		daip = luci.sys.exec("ip addr show 2>/dev/null | grep -w 'inet' | grep 'global' | grep 'brd' | grep -Eo 'inet [0-9\.]+' | awk '{print $2}' | head -n 1 | tr -d '\n'")
-	end
+	local daip
+	daip = fs.lanip()
 	return daip
 end
 
@@ -229,35 +203,31 @@ local function startlog()
 	return line_trans
 end
 
+local function pkg_type()
+	if fs.access("/usr/bin/apk") then
+		return "apk"
+	else
+		return "opkg"
+	end
+end
+
 local function coremodel()
 	if opkg and opkg.info("libc") and opkg.info("libc")["libc"] then
 		return opkg.info("libc")["libc"]["Architecture"]
 	else
-		return luci.sys.exec("rm -f /var/lock/opkg.lock && opkg status libc 2>/dev/null |grep 'Architecture' |awk -F ': ' '{print $2}' 2>/dev/null")
+		if fs.access("/bin/opkg") then
+			return luci.sys.exec("rm -f /var/lock/opkg.lock && opkg status libc 2>/dev/null |grep 'Architecture' |awk -F ': ' '{print $2}' 2>/dev/null")
+		elseif fs.access("/usr/bin/apk") then
+			return luci.sys.exec("apk list libc 2>/dev/null |awk '{print $2}'")
+		end
 	end
 end
 
 local function check_core()
-	if not nixio.fs.access(dev_core_path) and not nixio.fs.access(tun_core_path) and not nixio.fs.access(meta_core_path) then
+	if not nixio.fs.access(meta_core_path) then
 		return "0"
 	else
 		return "1"
-	end
-end
-
-local function corecv()
-	if not nixio.fs.access(dev_core_path) then
-		return "0"
-	else
-		return luci.sys.exec(string.format("%s -v 2>/dev/null |awk -F ' ' '{print $2}'", dev_core_path))
-	end
-end
-
-local function coretuncv()
-	if not nixio.fs.access(tun_core_path) then
-		return "0"
-	else
-		return luci.sys.exec(string.format("%s -v 2>/dev/null |awk -F ' ' '{print $2}'", tun_core_path))
 	end
 end
 
@@ -271,17 +241,19 @@ end
 
 local function corelv()
 	luci.sys.call("bash /usr/share/openclash/clash_version.sh")
-	local core_lv = luci.sys.exec("sed -n 1p /tmp/clash_last_version 2>/dev/null")
-	local core_tun_lv = luci.sys.exec("sed -n 2p /tmp/clash_last_version 2>/dev/null")
 	local core_meta_lv = luci.sys.exec("sed -n 3p /tmp/clash_last_version 2>/dev/null")
-	return core_lv .. "," .. core_tun_lv .. "," .. core_meta_lv
+	return core_meta_lv
 end
 
 local function opcv()
 	if opkg and opkg.info("luci-app-openclash") and opkg.info("luci-app-openclash")["luci-app-openclash"] then
 		return "v" .. opkg.info("luci-app-openclash")["luci-app-openclash"]["Version"]
 	else
-		return luci.sys.exec("rm -f /var/lock/opkg.lock && opkg status luci-app-openclash 2>/dev/null |grep 'Version' |awk -F 'Version: ' '{print \"v\"$2}'")
+		if fs.access("/bin/opkg") then
+			return luci.sys.exec("rm -f /var/lock/opkg.lock && opkg status luci-app-openclash 2>/dev/null |grep 'Version' |awk -F 'Version: ' '{print \"v\"$2}'")
+		elseif fs.access("/usr/bin/apk") then
+			return "v" .. luci.sys.exec("apk list luci-app-openclash 2>/dev/null |grep 'installed' | grep -oE '\\d+(\\.\\d+)*' | head -1")
+		end
 	end
 end
 
@@ -354,13 +326,9 @@ function core_download()
 	local cdn_url = luci.http.formvalue("url")
 	if cdn_url then
 		luci.sys.call(string.format("rm -rf /tmp/clash_last_version 2>/dev/null && bash /usr/share/openclash/clash_version.sh '%s' >/dev/null 2>&1", cdn_url))
-		luci.sys.call(string.format("bash /usr/share/openclash/openclash_core.sh 'Dev' '%s' >/dev/null 2>&1 &", cdn_url))
-		luci.sys.call(string.format("bash /usr/share/openclash/openclash_core.sh 'TUN' '%s' >/dev/null 2>&1 &", cdn_url))
 		luci.sys.call(string.format("bash /usr/share/openclash/openclash_core.sh 'Meta' '%s' >/dev/null 2>&1 &", cdn_url))
 	else
 		luci.sys.call("rm -rf /tmp/clash_last_version 2>/dev/null && bash /usr/share/openclash/clash_version.sh >/dev/null 2>&1")
-		luci.sys.call("bash /usr/share/openclash/openclash_core.sh 'Dev' >/dev/null 2>&1 &")
-		luci.sys.call("bash /usr/share/openclash/openclash_core.sh 'TUN' >/dev/null 2>&1 &")
 		luci.sys.call("bash /usr/share/openclash/openclash_core.sh 'Meta' >/dev/null 2>&1 &")
 	end
 
@@ -369,16 +337,6 @@ end
 function download_rule()
 	local filename = luci.http.formvalue("filename")
 	local state = luci.sys.call(string.format('/usr/share/openclash/openclash_download_rule_list.sh "%s" >/dev/null 2>&1',filename))
-	return state
-end
-
-function download_disney_domains()
-	local state = luci.sys.call(string.format('/usr/share/openclash/openclash_download_rule_list.sh "%s" >/dev/null 2>&1',"disney_domains"))
-	return state
-end
-
-function download_netflix_domains()
-	local state = luci.sys.call(string.format('/usr/share/openclash/openclash_download_rule_list.sh "%s" >/dev/null 2>&1',"netflix_domains"))
 	return state
 end
 
@@ -663,10 +621,18 @@ function set_subinfo_url()
 end
 
 function sub_info_get()
-	local filepath, filename, sub_url, sub_info, info, upload, download, total, expire, http_code, len, percent, day_left, day_expire, surplus, used
+	local sub_ua, filepath, filename, sub_url, sub_info, info, upload, download, total, expire, http_code, len, percent, day_left, day_expire, surplus, used
 	local info_tb = {}
 	filename = luci.http.formvalue("filename")
 	sub_info = ""
+	sub_ua = "Clash"
+	uci:foreach("openclash", "config_subscribe",
+		function(s)
+			if s.name == filename and s.sub_ua then
+				sub_ua = s.sub_ua
+			end
+		end
+	)
 	if filename and not is_start() then
 		uci:foreach("openclash", "subscribe_info",
 			function(s)
@@ -689,7 +655,7 @@ function sub_info_get()
 		if not sub_url then
 			sub_info = "No Sub Info Found"
 		else
-			info = luci.sys.exec(string.format("curl -sLI -X GET -m 10 -w 'http_code='%%{http_code} -H 'User-Agent: Clash' '%s'", sub_url))
+			info = luci.sys.exec(string.format("curl -sLI -X GET -m 10 -w 'http_code='%%{http_code} -H 'User-Agent: %s' '%s'", sub_ua, sub_url))
 			if not info or tonumber(string.sub(string.match(info, "http_code=%d+"), 11, -1)) ~= 200 then
 				info = luci.sys.exec(string.format("curl -sLI -X GET -m 10 -w 'http_code='%%{http_code} -H 'User-Agent: Quantumultx' '%s'", sub_url))
 			end
@@ -709,20 +675,34 @@ function sub_info_get()
 						expire = os.date("%Y-%m-%d", day_expire) or "null"
 						if day_expire and os.time() <= day_expire then
 							day_left = math.ceil((day_expire - os.time()) / (3600*24))
+							if math.ceil(day_left / 365) > 50 then
+								day_left = "∞"
+							end
 						elseif day_expire == nil then
 							day_left = "null"
 						else
 							day_left = 0
 						end
-						if used and total and used < total then
+						if used and total and used <= total then
 							percent = string.format("%.1f",((total-used)/total)*100) or nil
-						elseif used == nil or total == nil or total == 0 then
+							surplus = fs.filesize(total - used) or "null"
+						elseif used == nil and total and total > 0.0 then
 							percent = 100
+							surplus = total
+						elseif total and total == 0.0 then
+							percent = 100
+							surplus = "∞"
 						else
 							percent = 0
+							surplus = "null"
 						end
-						surplus = fs.filesize(total - used) or "null"
-						total = fs.filesize(total) or "null"
+						if total and total > 0.0 then
+							total = fs.filesize(total) or "null"
+						elseif total and total == 0.0 then
+							total = "∞"
+						else
+							total = "null"
+						end
 						used = fs.filesize(used) or "null"
 						sub_info = "Successful"
 					else
@@ -751,7 +731,7 @@ function action_rule_mode()
 		local daip = daip()
 		local dase = dase() or ""
 		local cn_port = cn_port()
-		core_type = uci:get("openclash", "config", "core_type") or "Dev"
+		core_type = uci:get("openclash", "config", "core_type") or "Meta"
 		if not daip or not cn_port then return end
 		info = json.parse(luci.sys.exec(string.format('curl -sL -m 3 -H "Content-Type: application/json" -H "Authorization: Bearer %s" -XGET http://"%s":"%s"/configs', dase, daip, cn_port)))
 		if info then
@@ -773,9 +753,7 @@ function action_switch_rule_mode()
 		local daip = daip()
 		local dase = dase() or ""
 		local cn_port = cn_port()
-		local core_type = uci:get("openclash", "config", "core_type") or "Dev"
 		mode = luci.http.formvalue("rule_mode")
-		if mode == script and core_type ~= "TUN" then luci.http.status(500, "Switch Faild") return end
 		if not daip or not cn_port then luci.http.status(500, "Switch Faild") return end
 		info = luci.sys.exec(string.format('curl -sL -m 3 -H "Content-Type: application/json" -H "Authorization: Bearer %s" -XPATCH http://"%s":"%s"/configs -d \'{\"mode\": \"%s\"}\'', dase, daip, cn_port, mode))
 		if info ~= "" then
@@ -867,7 +845,7 @@ end
 
 local function s(e)
 local t=0
-local a={' B/S',' KB/S',' MB/S',' GB/S',' TB/S'}
+local a={' B/S',' KB/S',' MB/S',' GB/S',' TB/S',' PB/S'}
 if (e<=1024) then
 	return e..a[1]
 else
@@ -881,15 +859,12 @@ end
 
 function action_toolbar_show_sys()
 	local pid = luci.sys.exec("pidof clash |head -1 |tr -d '\n' 2>/dev/null")
-	local mem, cpu
+	local cpu
 	if pid and pid ~= "" then
-		mem = tonumber(luci.sys.exec(string.format("cat /proc/%s/status 2>/dev/null |grep -w VmRSS |awk '{print $2}'", pid)))
 		cpu = luci.sys.exec(string.format("top -b -n1 |grep -E '%s' 2>/dev/null |grep -v grep |awk '{for (i=1;i<=NF;i++) {if ($i ~ /clash/) break; else cpu=i}}; {print $cpu}' 2>/dev/null", pid))
-		if mem and cpu then
-			mem = fs.filesize(mem*1024) or "0 KB"
+		if cpu then
 			cpu = string.match(cpu, "%d+") or "0"
 		else
-			mem = "0 KB"
 			cpu = "0"
 		end
 	else
@@ -897,7 +872,6 @@ function action_toolbar_show_sys()
 	end
 	luci.http.prepare_content("application/json")
 	luci.http.write_json({
-		mem = mem,
 		cpu = cpu;
 	})
 end
@@ -1081,8 +1055,7 @@ function action_status()
 		db_forward_ssl = db_foward_ssl(),
 		web = is_web(),
 		cn_port = cn_port(),
-		restricted_mode = restricted_mode(),
-		core_type = uci:get("openclash", "config", "core_type") or "Dev";
+		core_type = uci:get("openclash", "config", "core_type") or "Meta";
 	})
 end
 
@@ -1090,8 +1063,6 @@ function action_state()
 	luci.http.prepare_content("application/json")
 	luci.http.write_json({
 		lhie1 = lhie1(),
-		ConnersHua = ConnersHua(),
-		ConnersHua_return = ConnersHua_return(),
 		ipdb = ipdb(),
 		geosite = geosite(),
 		historychecktime = historychecktime(),
@@ -1117,8 +1088,6 @@ end
 function action_update()
 	luci.http.prepare_content("application/json")
 	luci.http.write_json({
-			corecv = corecv(),
-			coretuncv = coretuncv(),
 			coremetacv = coremetacv(),
 			coremodel = coremodel(),
 			opcv = opcv(),
@@ -1140,6 +1109,7 @@ function action_update_ma()
 	luci.http.prepare_content("application/json")
 	luci.http.write_json({
 			oplv = oplv(),
+			pkg_type = pkg_type(),
 			corelv = corelv(),
 			corever = corever();
 	})
@@ -1195,20 +1165,6 @@ function action_download_rule()
 	luci.http.prepare_content("application/json")
 	luci.http.write_json({
 		rule_download_status = download_rule();
-	})
-end
-
-function action_download_netflix_domains()
-	luci.http.prepare_content("application/json")
-	luci.http.write_json({
-		rule_download_status = download_netflix_domains();
-	})
-end
-
-function action_download_disney_domains()
-	luci.http.prepare_content("application/json")
-	luci.http.write_json({
-		rule_download_status = download_disney_domains();
 	})
 end
 
@@ -1326,38 +1282,6 @@ function split(str,delimiter)
 	return arr
 end
 
-function action_write_netflix_domains()
-	local domains = luci.http.formvalue("domains")
-	local dustom_file = "/etc/openclash/custom/openclash_custom_netflix_domains.list"
-	local file = io.open(dustom_file, "a+")
-	file:seek("set")
-	local domain = file:read("*a")
-	for v, k in pairs(split(domains,"\n")) do
-		if not string.find(domain,k,1,true) then
-			file:write(k.."\n")
-		end
-	end
-	file:close()
-	return
-end
-
-function action_catch_netflix_domains()
-	local cmd = "/usr/share/openclash/openclash_debug_getcon.lua 'netflix-nflxvideo'"
-	luci.http.prepare_content("text/plain")
-	local util = io.popen(cmd)
-	if util and util ~= "" then
-		while true do
-			local ln = util:read("*l")
-			if not ln then break end
-			luci.http.write(ln)
-			luci.http.write(",")
-		end
-		util:close()
-		return
-	end
-	luci.http.status(500, "Bad address")
-end
-
 function action_diag_connection()
 	local addr = luci.http.formvalue("addr")
 	if addr and (datatype.hostname(addr) or datatype.ipaddr(addr)) then
@@ -1425,8 +1349,8 @@ function action_backup()
 	local reader = ltn12_popen("tar -C '/etc/openclash/' -cz . 2>/dev/null")
 
 	luci.http.header(
-		'Content-Disposition', 'attachment; filename="Backup-OpenClash-%s.tar.gz"' %{
-			os.date("%Y-%m-%d-%H-%M-%S")
+		'Content-Disposition', 'attachment; filename="Backup-OpenClash-%s-%s-%s.tar.gz"' %{
+			device_name, device_arh, os.date("%Y-%m-%d-%H-%M-%S")
 		})
 
 	luci.http.prepare_content("application/x-targz")
@@ -1439,8 +1363,8 @@ function action_backup_ex_core()
 	local reader = ltn12_popen("echo 'core' > /tmp/oc_exclude.txt && tar -C '/etc/openclash/' -X '/tmp/oc_exclude.txt' -cz . 2>/dev/null")
 
 	luci.http.header(
-		'Content-Disposition', 'attachment; filename="Backup-OpenClash-Exclude-Cores-%s.tar.gz"' %{
-			os.date("%Y-%m-%d-%H-%M-%S")
+		'Content-Disposition', 'attachment; filename="Backup-OpenClash-Exclude-Cores-%s-%s-%s.tar.gz"' %{
+			device_name, device_arh, os.date("%Y-%m-%d-%H-%M-%S")
 		})
 
 	luci.http.prepare_content("application/x-targz")
@@ -1452,8 +1376,8 @@ function action_backup_only_config()
 	local reader = ltn12_popen("tar -C '/etc/openclash' -cz './config' 2>/dev/null")
 
 	luci.http.header(
-		'Content-Disposition', 'attachment; filename="Backup-OpenClash-Config-%s.tar.gz"' %{
-			os.date("%Y-%m-%d-%H-%M-%S")
+		'Content-Disposition', 'attachment; filename="Backup-OpenClash-Config-%s-%s-%s.tar.gz"' %{
+			device_name, device_arh, os.date("%Y-%m-%d-%H-%M-%S")
 		})
 
 	luci.http.prepare_content("application/x-targz")
@@ -1464,8 +1388,8 @@ function action_backup_only_core()
 	local reader = ltn12_popen("tar -C '/etc/openclash' -cz './core' 2>/dev/null")
 
 	luci.http.header(
-		'Content-Disposition', 'attachment; filename="Backup-OpenClash-Cores-%s.tar.gz"' %{
-			os.date("%Y-%m-%d-%H-%M-%S")
+		'Content-Disposition', 'attachment; filename="Backup-OpenClash-Cores-%s-%s-%s.tar.gz"' %{
+			device_name, device_arh, os.date("%Y-%m-%d-%H-%M-%S")
 		})
 
 	luci.http.prepare_content("application/x-targz")
@@ -1476,8 +1400,8 @@ function action_backup_only_rule()
 	local reader = ltn12_popen("tar -C '/etc/openclash' -cz './rule_provider' 2>/dev/null")
 
 	luci.http.header(
-		'Content-Disposition', 'attachment; filename="Backup-OpenClash-Only-Rule-Provider-%s.tar.gz"' %{
-			os.date("%Y-%m-%d-%H-%M-%S")
+		'Content-Disposition', 'attachment; filename="Backup-OpenClash-Only-Rule-Provider-%s-%s-%s.tar.gz"' %{
+			device_name, device_arh, os.date("%Y-%m-%d-%H-%M-%S")
 		})
 
 	luci.http.prepare_content("application/x-targz")
@@ -1488,8 +1412,8 @@ function action_backup_only_proxy()
 	local reader = ltn12_popen("tar -C '/etc/openclash' -cz './proxy_provider' 2>/dev/null")
 
 	luci.http.header(
-		'Content-Disposition', 'attachment; filename="Backup-OpenClash-Proxy-Provider-%s.tar.gz"' %{
-			os.date("%Y-%m-%d-%H-%M-%S")
+		'Content-Disposition', 'attachment; filename="Backup-OpenClash-Proxy-Provider-%s-%s-%s.tar.gz"' %{
+			device_name, device_arh, os.date("%Y-%m-%d-%H-%M-%S")
 		})
 
 	luci.http.prepare_content("application/x-targz")
